Added excludeBoundariesSize argument to identifyMajorPeaks().
Before, nearbyWinSize was used for two different but related criteria: the
range for including peaks close to a large peak AND the range to
exclude peaks close to the beginning and end of the signal. Now,
we have two independent arguments for each setting.
The current behaviour does not change, but it is now more
flexible. Related to #4.